fix: revert module import for tosca interpreter
Module import in calculate_derivatives was rewritten for unittest but
broke execution of the workflow within the optimization environment as
the package can not be installed into the Tosca Python interpreter.

BREAKING CHANGE: Unittest is broken for now as the included utils module
can not be imported from calculate_derivatives without referencing the
package when running the unittest.